--- src/services/aiProvider/type.ts
@@ -1,27 +1,0 @@
-import {
-  AiProviderDetailItem,
-  AiProviderRuntimeState,
-  AiProviderSortMap,
-  CreateAiProviderParams,
-  UpdateAiProviderConfigParams,
-} from '@/types/aiProvider';
-
-export interface IAiProviderService {
-  createAiProvider: (params: CreateAiProviderParams) => Promise<any>;
-
-  deleteAiProvider: (id: string) => Promise<any>;
-
-  getAiProviderById: (id: string) => Promise<AiProviderDetailItem | undefined>;
-
-  getAiProviderList: () => Promise<any>;
-
-  getAiProviderRuntimeState: (isLogin?: boolean) => Promise<AiProviderRuntimeState>;
-
-  toggleProviderEnabled: (id: string, enabled: boolean) => Promise<any>;
-
-  updateAiProvider: (id: string, value: any) => Promise<any>;
-
-  updateAiProviderConfig: (id: string, value: UpdateAiProviderConfigParams) => Promise<any>;
-
-  updateAiProviderOrder: (items: AiProviderSortMap[]) => Promise<any>;
-}
