--- src/services/export/type.ts
@@ -1,5 +1,0 @@
-import { ExportDatabaseData } from '@/types/export';
-
-export interface IExportService {
-  exportData(): Promise<ExportDatabaseData>;
-}
