--- src/services/message/type.ts
@@ -1,55 +1,0 @@
-import {
-  ChatMessageError,
-  ChatMessagePluginError,
-  ChatTTS,
-  ChatTranslate,
-  CreateMessageParams,
-  CreateMessageResult,
-  ModelRankItem,
-  UIChatMessage,
-  UpdateMessageParams,
-  UpdateMessageRAGParams,
-  UpdateMessageResult,
-} from '@lobechat/types';
-import type { HeatmapsProps } from '@lobehub/charts';
-
-/* eslint-disable typescript-sort-keys/interface */
-
-export interface IMessageService {
-  createMessage(data: CreateMessageParams): Promise<string>;
-  createNewMessage(data: CreateMessageParams): Promise<CreateMessageResult>;
-
-  getMessages(sessionId: string, topicId?: string, groupId?: string): Promise<UIChatMessage[]>;
-  getGroupMessages(groupId: string, topicId?: string): Promise<UIChatMessage[]>;
-  countMessages(params?: {
-    endDate?: string;
-    range?: [string, string];
-    startDate?: string;
-  }): Promise<number>;
-  countWords(params?: {
-    endDate?: string;
-    range?: [string, string];
-    startDate?: string;
-  }): Promise<number>;
-  rankModels(): Promise<ModelRankItem[]>;
-  getHeatmaps(): Promise<HeatmapsProps['data']>;
-  updateMessageError(id: string, error: ChatMessageError): Promise<any>;
-  updateMessage(
-    id: string,
-    message: Partial<UpdateMessageParams>,
-    options?: { sessionId?: string | null; topicId?: string | null },
-  ): Promise<UpdateMessageResult>;
-  updateMessageTTS(id: string, tts: Partial<ChatTTS> | false): Promise<any>;
-  updateMessageTranslate(id: string, translate: Partial<ChatTranslate> | false): Promise<any>;
-  updateMessagePluginState(id: string, value: Record<string, any>): Promise<any>;
-  updateMessagePluginError(id: string, value: ChatMessagePluginError | null): Promise<any>;
-  updateMessageRAG(id: string, value: UpdateMessageRAGParams): Promise<void>;
-  updateMessagePluginArguments(id: string, value: string | Record<string, any>): Promise<any>;
-  removeMessage(id: string): Promise<any>;
-  removeMessages(ids: string[]): Promise<any>;
-  removeMessagesByAssistant(assistantId: string, topicId?: string): Promise<any>;
-  removeMessagesByGroup(groupId: string, topicId?: string): Promise<any>;
-  removeAllMessages(): Promise<any>;
-  messageCountToCheckTrace(): Promise<boolean>;
-  hasMessages(): Promise<boolean>;
-}
